@@ -6,11 +6,11 @@
     "homepage": "https://github.com/twilio/authy-php",
     "license": "MIT",
     "require": {
-        "php": ">=5.6.0",
-        "guzzlehttp/guzzle": "~6.0"
+        "php": "^7.2.5",
+        "guzzlehttp/guzzle": "~7.0"
     },
     "require-dev": {
-        "EHER/PHPUnit": ">= 1.6"
+        "phpunit/phpunit": "^8.0"
     },
     "authors": [
         {
